Add files via upload
A full stack health monitoring system which can be used as an add-on for any hospital management system to track and monitor the health of its patients.

In this project, we worked on health monitoriting system which keeps track of diseases present in the society. It helps people with their appointments with the hospital. People can gain knowledge about the diseases present in society. It can help create awareness in the government and also people towards the health issues the present society is facing. Based on this, the government can be able to decide on which medicines they can spend the resources on. Visualization was also done for people to easily understand about the disease info. This helps us to highlight the trend or useful information.